Solve the problem of writing Chinese garbled code to csv training set files
When reading csv files, Excel identifies the encoding by reading the BOM on the file header. If the file header has no BOM information, it will read according to the Unicode encoding by default. The default csv is ANSI. When we use utf-8 encoding to generate a csv file, no BOM information is generated, and Excel will automatically read it according to the Unicode encoding, resulting in garbled code problems.